+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 ## next release
 
+- Ability to parametrize machines order on the booking page
 - Fix a bug: rake task fablab:fix:categories_slugs bash interpretation error
 - Fix a bug: file inputs filled with long filenames render improperly with an overflow
 - Fix a bug: title concordance radio buttons render improperly on smaller screens
 - [TODO DEPLOY] `rake fablab:fix:categories_slugs`
+- [TODO DEPLOY] `rake db:seed`
 
 ## v2.6.5 2018 July 24